chore: split DefaultScrollbars.ts into files
