+++ game/game.json
@@ -14,14 +14,24 @@
 			"path": "script/Scrollable.js",
 			"global": true
 		},
-		"DefaultScrollbars": {
-			"type": "script",
-			"path": "script/DefaultScrollbars.js",
-			"global": true
-		},
 		"createDefaultScrollbarImage": {
 			"type": "script",
 			"path": "script/createDefaultScrollbarImage.js",
+			"global": true
+		},
+		"DefaultVerticalScrollbar": {
+			"type": "script",
+			"path": "script/DefaultVerticalScrollbar.js",
+			"global": true
+		},
+		"NullScrollbar": {
+			"type": "script",
+			"path": "script/NullScrollbar.js",
+			"global": true
+		},
+		"ScrollbarLike": {
+			"type": "script",
+			"path": "script/ScrollbarLike.js",
 			"global": true
 		}
 	},
